+++ components/common/src/components/nv-checkbox/index.js
@@ -1,0 +1,16 @@
+/**
+ * NoahV
+ * Copyright (c) 2016 Baidu, Inc. All Rights Reserved.
+ *
+ * @file NvCheckbox Component
+ * @author nimingdexiaohai(nimingdexiaohai@163.com)
+ */
+
+import NvCheckbox from './checkbox.vue';
+
+export {
+    NvCheckbox
+};
+
+export default NvCheckbox;
+